fix(app): Fuenf Befunde aus der Bisektion der letzten vier Kalibrier-Presets —           vom Zufallszug zuviel bis zur Bildspalte, die ein Movement einsammelt.
Channel Shift zog die Zufalls-Permutation schon im ersten Frame
(`apeChanMode < 0`); die Referenz zieht NUR auf dem Beat und zeigt bis dahin die
Vorgabe (r_chanshift:124-126). Ein Zug zuviel aus dem geteilten Strom — danach
war jede weitere Permutation versetzt (Referenz BGR, wir BRG). Der Startwert
kommt jetzt aus dem PRESET-Wert und wird bei Aenderung neu gesetzt, wie
`interferences.rotation` seit S57.

Convolution rechnet in der APE ganzzahlig in BYTES:
`x = (summe + bias*256) / scale`, Richtung Null abgeschnitten, dann 0..255
geklemmt. `bias` zaehlt also in ganzen 256ern — schon bias=1 hebt jeden Kanal
ueber 255; wir addierten bias/255. Belegt an (200,100,40): bias=1/scale=2 gibt
(228,178,148), bias=1/scale=256 gibt (1,1,1). Dazu: `absolute` liefert bei
negativem Ergebnis 255 statt des Betrags, `twoPass` VERDOPPELT statt zweimal zu
falten, `scale=0` gilt als 1, und der Wert von `bias` wirkt als
vorzeichenbehaftetes Byte. 33 Sonden gegen AvsRef treffen jetzt exakt.

Die 50/50-Mischung ist die abgeschnittene Byte-Mittelung der Referenz
(`floor(a/2)+floor(b/2)`, r_defs.h:504-544), nicht `(a+b)/2`.

Picture II: der Dateiname steht in einem FESTEN 260-Byte-Feld (MAX_PATH), nicht
als NUL-terminierte Zeichenkette — der Blob misst 284 = 260 + 6*4. Ein variabler
Lauf endet nach dem ersten NUL und liest den Rest des Puffers als Zahlen; bei
"The Real Impressionist" stand dort noch ".bmp" von einer frueheren Eingabe und
alle sechs Felder waren Muell (blendMode 1886350382 statt 0). Das Bild kam
damit nie an: 0,516 -> 0,173.

SuperScope-Punkte sind jetzt DOUBLE. `x=1-2/w` ergibt in double exakt Spalte
w-1, ueber float gerundet w-1-eps und damit w-2. Die Formel in
renderPixelDots war richtig, der Eingangswert war verloren. Der Fehler steckt in
jeder Bildbreite, faellt aber nur auf, wenn ein Effekt genau diese Spalte
abtastet — in "The Real Impressionist" sammelt ein Movement bei 640x480 alle
Ausgabepixel dort ein. Die Vertex-Puffer bekommen weiter float, jetzt
ausdruecklich gecastet.

Dynamic Shift ganz nach r_shift.cpp:190-310: ganzzahliger Anteil plus 8-Bit-
Bruchteil mit dem asymmetrischen Kipp-Zweig (`else { ++a; part = 255-part; }`),
freigewordene Zeilen und Spalten HART auf Schwarz, BLEND4 mit
Gewichtssumme 255, und der Blend in der richtigen Richtung
(`BLEND_ADJ(verschoben, framebuffer, ialpha)`; ialpha<=0 heisst Rueckkehr, nicht
Durchsicht). Schon eine Verschiebung von null in y kostet dort die erste und
letzte Zeile — genau der Saum, der bei uns stehen blieb.

Gemessen, je 240 Frames gegen AvsRef, in zwei Groessen:
  Alternate Reality      0,622/0,539 -> 0,136/0,339  ·  0,867/0,638 -> 0,077/0,230
  Picture II             0,516/0,488 -> 0,173/0,180  ·  0,450/0,402 -> 0,235/0,249
  Rotor                  0,461       -> 0,018/0,020  ·               -> 0,006/0,016  GRUEN
  Bright Light District  0,252/0,234 -> 0,026/0,162  ·  0,206/0,186 -> 0,066/0,110
Matrix 40/43 unveraendert, Modul-Sonden 83/87, Tests 485/485 gruen, 0 Skips.

+++ projects/apps/MyViz/include/visualizers/modules/SuperscopeModule.hpp
@@ -111,8 +111,15 @@ enum class SuperscopePreset
  */
 struct SuperscopePoint
 {
-    float x = 0.0f;         ///< X coordinate (-1 to 1)
-    float y = 0.0f;         ///< Y coordinate (-1 to 1)
+    /// X/Y in NDC (-1..1) als DOUBLE — wie in AVS. Der Zielpixel entsteht dort
+    /// aus `(int)((x+1)*w*0.5)`, und das ist ein Kantenfall: `x=1-2/w` ergibt in
+    /// double exakt `w-1`, ueber float gerundet aber `w-1-eps` und damit `w-2`.
+    /// Wir schrieben die Skriptwerte frueher als float weg und verloren die
+    /// letzte Spalte (Befund S58: bei "The Real Impressionist" faengt ein
+    /// Movement genau diese Spalte ein und zieht sie ueber das ganze Bild).
+    /// Die Vertex-Puffer bekommen weiterhin float — dort ist es GL-Genauigkeit.
+    double x = 0.0;         ///< X coordinate (-1 to 1)
+    double y = 0.0;         ///< Y coordinate (-1 to 1)
     float r = 1.0f;         ///< Red (0 to 1)
     float g = 1.0f;         ///< Green (0 to 1)
     float b = 1.0f;         ///< Blue (0 to 1)
